backend/requirements.txt
Django==5.2.7
djangorestframework==3.16.1
djoser==2.3.3
django-filter==25.2
Pillow==12.0.0
gunicorn==23.0.0
psycopg2-binary==2.9.11
flake8==6.0.0